+++ main.go
@@ -13,8 +13,7 @@ import (
 
 	"github.com/urfave/cli/v2"
 
-	"github.com/jlegrone/tctx/config/config"
-	"github.com/jlegrone/tctx/internal/configrw"
+	"github.com/jlegrone/tctx/config"
 
 	"github.com/jlegrone/tctx/internal/xbar"
 )
@@ -156,12 +155,12 @@ func parseAdditionalEnvVars(input []string) (additional map[string]string, err e
 	return envVars, nil
 }
 
-func switchContexts(w io.Writer, rw *configrw.FSReaderWriter, contextName, namespace string) error {
-	if err := rw.SetActiveContext(contextName, namespace); err != nil {
+func switchContexts(w io.Writer, t *config.Tctx, contextName, namespace string) error {
+	if err := t.SetActiveContext(contextName, namespace); err != nil {
 		return err
 	}
 
-	cfg, err := rw.GetContext(contextName)
+	cfg, err := t.GetContext(contextName)
 	if err != nil {
 		return err
 	}
@@ -195,22 +194,22 @@ func newApp(configFile string) *cli.App {
 						return err
 					}
 
-					rw, err := configrw.NewReaderWriter(path)
+					t, err := config.NewTctx(config.WithConfigFile(path))
 					if err != nil {
 						return err
 					}
 
 					// Error if context already exists
-					existingCfg, _ := rw.GetContext(name)
+					existingCfg, _ := t.GetContext(name)
 					if existingCfg != nil {
 						return fmt.Errorf("a context with name %q already exists", name)
 					}
 
-					if err := rw.UpsertContext(name, cfg); err != nil {
+					if err := t.UpsertContext(name, cfg); err != nil {
 						return err
 					}
 
-					return switchContexts(c.App.Writer, rw, name, cfg.Namespace)
+					return switchContexts(c.App.Writer, t, name, cfg.Namespace)
 				},
 			},
 			{
@@ -223,21 +222,21 @@ func newApp(configFile string) *cli.App {
 						return err
 					}
 
-					rw, err := configrw.NewReaderWriter(path)
+					t, err := config.NewTctx(config.WithConfigFile(path))
 					if err != nil {
 						return err
 					}
 
 					// Check that context already exists
-					if _, err := rw.GetContext(name); err != nil {
+					if _, err := t.GetContext(name); err != nil {
 						return err
 					}
 
-					if err := rw.UpsertContext(name, newCfg); err != nil {
+					if err := t.UpsertContext(name, newCfg); err != nil {
 						return err
 					}
 
-					return switchContexts(c.App.Writer, rw, name, newCfg.Namespace)
+					return switchContexts(c.App.Writer, t, name, newCfg.Namespace)
 				},
 			},
 			{
@@ -248,13 +247,13 @@ func newApp(configFile string) *cli.App {
 					getContextFlag(true),
 				},
 				Action: func(c *cli.Context) error {
-					rw, err := configrw.NewReaderWriter(c.String(configPathFlag))
+					t, err := config.NewTctx(config.WithConfigFile(c.String(configPathFlag)))
 					if err != nil {
 						return err
 					}
 
 					contextName := c.String(contextNameFlag)
-					if err := rw.DeleteContext(contextName); err != nil {
+					if err := t.DeleteContext(contextName); err != nil {
 						return err
 					}
 
@@ -268,11 +267,11 @@ func newApp(configFile string) *cli.App {
 				Aliases: []string{"ls"},
 				Usage:   "list contexts",
 				Action: func(c *cli.Context) error {
-					rw, err := configrw.NewReaderWriter(c.String(configPathFlag))
+					t, err := config.NewTctx(config.WithConfigFile(c.String(configPathFlag)))
 					if err != nil {
 						return err
 					}
-					contexts, err := rw.GetAllContexts()
+					contexts, err := t.GetAllContexts()
 					if err != nil {
 						return err
 					}
@@ -314,12 +313,12 @@ func newApp(configFile string) *cli.App {
 						namespace   = c.String(namespaceFlag)
 					)
 
-					rw, err := configrw.NewReaderWriter(configPath)
+					t, err := config.NewTctx(config.WithConfigFile(configPath))
 					if err != nil {
 						return err
 					}
 
-					return switchContexts(c.App.Writer, rw, contextName, namespace)
+					return switchContexts(c.App.Writer, t, contextName, namespace)
 				},
 			},
 			{
@@ -335,11 +334,11 @@ func newApp(configFile string) *cli.App {
 						return err
 					}
 
-					rw, err := configrw.NewReaderWriter(c.String(configPathFlag))
+					t, err := config.NewTctx(config.WithConfigFile(c.String(configPathFlag)))
 					if err != nil {
 						return err
 					}
-					cfg, err := rw.GetAllContexts()
+					cfg, err := t.GetAllContexts()
 					if err != nil {
 						return err
 					}
@@ -371,20 +370,20 @@ func newApp(configFile string) *cli.App {
 						return cli.ShowCommandHelp(c, "exec")
 					}
 
-					rw, err := configrw.NewReaderWriter(c.String(configPathFlag))
+					t, err := config.NewTctx(config.WithConfigFile(c.String(configPathFlag)))
 					if err != nil {
 						return err
 					}
 
 					contextName := c.String(contextNameFlag)
 					if contextName == "" {
-						contextName, err = rw.GetActiveContextName()
+						contextName, err = t.GetActiveContextName()
 						if err != nil {
 							return err
 						}
 					}
 
-					cfg, err := rw.GetContext(contextName)
+					cfg, err := t.GetContext(contextName)
 					if err != nil {
 						return err
 					}
